requirements.txt
Flask~=1.1.2
Flask-Cors==3.0.8
transformers~=2.11.0
tensorflow~=2.2.0
nltk~=3.5
google-cloud-storage~=1.28.1
google-cloud-datastore==1.12.0
pdfminer.six~=20200517
sklearn~=0.0
scikit-learn~=0.23.1
numpy~=1.18.5